+++ Hyejun/stack.cpp
@@ -1,0 +1,48 @@
+#include <iostream>
+#include "deque.h"
+
+class myStack{
+	private:
+		myDeque d;
+
+	public:
+		myStack()
+		{
+		
+		}
+
+		~myStack()
+		{
+
+		}
+
+		bool empty()
+		{
+			return d.empty();
+		}
+
+		int size()
+		{
+			return d.size();
+		}
+
+		int top()
+		{
+			return d.back();
+		}
+
+		void push(int data)
+		{
+			d.push_back(data);
+		}
+
+		void pop()
+		{
+			d.pop_back();
+		}
+};
+
+int main(void)
+{
+	return 0;
+}
